chore(tests): refactor tests into per-year files
No functional change; just restructuring.
